requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwcrypto==1.5.6
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
